ingredients can now be deleted from the ingredient list. Some parts of the code cleaned up.
