+++ src/main/data/ingredients.txt
@@ -1,2 +1,1 @@
-apple | 5
-apple | 5
+prawn | 5
