+++ src/main/java/duke/common/IngredientMessages.java
@@ -7,6 +7,7 @@ public class IngredientMessages {
 
     public static final String MESSAGE_INGREDIENT_DELETED = "     Noted. I've removed this ingredient:\n";
     public static final String MESSAGE_INGREDIENT_ADDED = "     Got it. I've added this ingredient:\n";
+    public static final String MESSAGE_LIST_INGREDIENTS = "     Here are the ingredients in your inventory:";
 
     public static final String ERROR_MESSAGE_INCOMPLETE = "     Please provide both the ingredient name and quantity.";
     public static final String ERROR_MESSAGE_ADD_INCORRECT_FORMAT = "     Please input the command in the following format:\n"
